feat: player object and main components
Game class defines the main game object, which
will handle the events and rendering.

Player class is the main player through which
user will play the game.

TextureManager is used for drawing on the renderer.

Watermelon and Lassi are entities which the player
can interact with. They inherit from the GameEntity
interface.